requirements.txt
aioredis==1.2.0
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
appdirs==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
astroid==1.6.5
astropy==3.0.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
black==18.9b0
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
bootstrap4==0.0.5
boto==2.48.0
boto3==1.9.104
botocore==1.12.104
Bottleneck==1.2.1
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
channels==2.1.6
channels-redis==2.3.3
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.3.9
coloredlogs==10.0
comtypes==1.1.4
conda==4.5.8
conda-build==3.10.5
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
cycler==0.10.0
cymem==2.0.2
Cython==0.28.2
cytoolz==0.9.0.1
daphne==2.2.5
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.9
distributed==1.21.8
Django==2.1.4
django-bootstrap==0.2.4
django-braces==1.13.0
django-channels==0.7.0
django-mangopay==1.0.0
django-model-utils==3.1.2
django-paypal==0.2.5
django-private-chat==0.2.2
django-templated-mail==1.1.1
django-uni-form==0.9.0
djangorestframework==3.9.0
djoser==1.4.0
docutils==0.14
drf-yasg==1.12.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
future==0.17.1
gevent==1.2.2
glob2==0.6
greenlet==0.4.13
h5py==2.7.1
heapdict==1.0.0
hiredis==1.0.0
html5lib==1.0.1
humanfriendly==4.18
hyperlink==17.3.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
incremental==17.5.0
inflection==0.3.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kiwisolver==1.0.1
klein==17.10.0
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lockfile==0.12.2
lxml==4.2.1
mangopaysdk==0.311
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.6.0
msgpack-numpy==0.4.3.2
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==1.0.2
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.5
numpydoc==0.8.0
oauthlib==3.0.1
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==18.0
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plac==0.9.6
pluggy==0.6.0
ply==3.11
preshed==2.0.1
prompt-toolkit==1.0.15
psutil==5.4.5
psycopg2-binary==2.7.5
py==1.5.3
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==1.8.4
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypiwin32==223
pyreadline==2.1
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-crfsuite==0.9.6
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.7
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
qrcode==6.1
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
razorpay==1.1.1
regex==2018.1.10
requests==2.20.0
requests-oauthlib==1.2.0
rope==0.10.7
ruamel-yaml==0.15.35
ruamel.yaml==0.15.78
s3transfer==0.2.0
scikit-image==0.13.1
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sortedcollections==0.6.1
sortedcontainers==1.5.10
South==1.0.2
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.3
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toml==0.10.0
toolz==0.9.0
tornado==5.0.2
tqdm==4.19.5
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websockets==3.2
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zope.interface==4.6.0
